pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "airflow-lite"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "pyarrow>=14.0",
    "pandas>=2.1",
    "tenacity>=8.2",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
phase2 = [
    "apscheduler>=3.10",
    "fastapi>=0.110",
    "uvicorn>=0.27",
    "pydantic>=2.6",
    "httpx>=0.27",
    "pywin32>=306",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/airflow_lite"]

[tool.pytest.ini_options]
testpaths = ["tests"]
